service/start: check check-migrations failure more robustly (#888)
If check-migrations fails with non-1 exit code, the server will continue to start.

The existing check-migrations call and test for exit code 1 was originally introduced as a fix for #461.  The change in this PR should be more robust than the existing approach, but `set -e` and `set -o pipefail` might provide even more safety in handling failures in this script.  These latter options would also allow for complete removal of `make check-migrations` from production.

There may also be other critical scripts in this repo which would benefit from similar changes.